Delete the users app, make changes to the code
