+++ api_yamdb/reviews/models.py
@@ -1,9 +1,42 @@
+from django.contrib.auth.models import AbstractUser
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
 from django.utils import timezone
 
 from api_yamdb.settings import LENGTH_STR
-from users.models import User
+# from users.models import User
+
+
+UUSER = 'user'
+MODERATOR = 'moderator'
+ADMIN = 'admin'
+ROLE_CHOICES = [
+    (UUSER, 'user'),
+    (MODERATOR, 'moderator'),
+    (ADMIN, 'admin')
+]
+
+
+class User(AbstractUser):
+    email = models.EmailField(unique=True)
+    bio = models.TextField(blank=True)
+    role = models.CharField(
+        max_length=20,
+        choices=ROLE_CHOICES,
+        default=UUSER
+    )
+    confirmation_code = models.CharField(
+        max_length=255,
+        blank=True,
+    )
+
+    @property
+    def is_admin(self):
+        return self.role == ADMIN or self.is_superuser
+
+    @property
+    def is_moderator(self):
+        return self.role == MODERATOR
 
 
 class Category(models.Model):
