--- api_yamdb/users/models.py
@@ -1,33 +1,0 @@
-from django.contrib.auth.models import AbstractUser
-from django.db import models
-
-USER = 'user'
-MODERATOR = 'moderator'
-ADMIN = 'admin'
-ROLE_CHOICES = [
-    (USER, 'user'),
-    (MODERATOR, 'moderator'),
-    (ADMIN, 'admin')
-]
-
-
-class User(AbstractUser):
-    email = models.EmailField(unique=True)
-    bio = models.TextField(blank=True)
-    role = models.CharField(
-        max_length=20,
-        choices=ROLE_CHOICES,
-        default=USER
-    )
-    confirmation_code = models.CharField(
-        max_length=255,
-        blank=True,
-    )
-
-    @property
-    def is_admin(self):
-        return self.role == ADMIN or self.is_superuser
-
-    @property
-    def is_moderator(self):
-        return self.role == MODERATOR
